+++ R/qsys_lsf.r
@@ -11,10 +11,18 @@ LSF = R6::R6Class("LSF",
             super$initialize(..., template=template)
         },
 
-        submit_jobs = function(...) {
-            opts = private$fill_options(...)
+        submit_jobs = function(n_jobs, ..., log_worker=FALSE, verbose=TRUE) {
+            opts = private$fill_options(n_jobs=n_jobs, ...)
             private$job_id = opts$job_name
-            filled = private$fill_template(opts)
+            if (!is.null(opts$log_file))
+                opts$log_file = normalizePath(opts$log_file, mustWork=FALSE)
+            else if (log_worker)
+                opts$log_file = paste0(private$job_id, "-%I.log")
+            filled = fill_template(private$template, opts,
+                                   required=c("master", "job_name", "n_jobs"))
+
+            if (verbose)
+                message("Submitting ", n_jobs, " worker jobs (ID: ", private$job_id, ") ...")
 
             success = system("bsub", input=filled, ignore.stdout=TRUE)
             if (success != 0) {