+++ R/qsys_ssh.r
@@ -15,32 +15,31 @@ SSH = R6::R6Class("SSH",
                 stop("Option 'clustermq.ssh.host' required for SSH but not set")
 
             super$initialize(..., template=template)
-            private$proxy_socket = rzmq::init.socket(private$zmq_context, "ZMQ_REP")
 
             # set forward and run ssh.r (send port, master)
             opts = private$fill_options(ssh_log=ssh_log, ssh_host=ssh_host)
-            ssh_cmd = private$fill_template(opts)
+            ssh_cmd = fill_template(private$template, opts,
+                required=c("ctl_port", "local_port", "job_port", "fwd_port", "ssh_host"))
 
             # wait for ssh to connect
             message(sprintf("Connecting %s via SSH ...", ssh_host))
             system(ssh_cmd, wait=TRUE, ignore.stdout=TRUE, ignore.stderr=TRUE)
 
             # Exchange init messages with proxy
-            init_timeout = getOption("clustermq.ssh.timeout", 5)
-            poll = rzmq::poll.socket(list(private$proxy_socket), list("read"),
-                                     timeout=init_timeout)
-            if (!poll[[1]]$read)
+            init_timeout = getOption("clustermq.ssh.timeout", 5) * 1000
+            answer = private$zmq$poll("proxy", timeout=init_timeout)
+            if (!answer)
                 stop("Remote R process did not respond after ",
                      init_timeout, " seconds. ",
                      "Check your SSH server log.")
-            msg = rzmq::receive.socket(private$proxy_socket)
+            msg = private$zmq$receive("proxy")
             if (msg$id != "PROXY_UP")
                 stop("Expected PROXY_UP, received ", sQuote(msg$id))
 
             # send common data to ssh
             message("Sending common data ...")
-            rzmq::send.socket(private$proxy_socket, data=c(list(id="DO_SETUP"), data))
-            msg = rzmq::receive.socket(private$proxy_socket)
+            private$zmq$send(c(list(id="DO_SETUP"), data), "proxy")
+            msg = private$zmq$receive("proxy")
             if (msg$id != "PROXY_READY")
                 stop("Expected PROXY_READY, received ", sQuote(msg$id))
 
@@ -48,7 +47,7 @@ SSH = R6::R6Class("SSH",
                                  token = msg$token)
         },
 
-        submit_jobs = function(...) {
+        submit_jobs = function(..., verbose=TRUE) {
             if (is.null(private$master))
                 stop("Need to call listen_socket() first")
 
@@ -64,10 +63,9 @@ SSH = R6::R6Class("SSH",
             # forward the submit_job call via ssh
             call[[1]] = quote(qsys$submit_jobs) #FIXME: only works bc 'qsys' in ssh_proxy
             call[2:length(call)] = evaluated
-            rzmq::send.socket(private$proxy_socket,
-                              data = list(id="PROXY_CMD", exec=call))
+            private$zmq$send(list(id="PROXY_CMD", exec=call), "proxy")
 
-            msg = rzmq::receive.socket(private$proxy_socket)
+            msg = private$zmq$receive("proxy")
             if (msg$id != "PROXY_CMD" || class(msg$reply) == "try-error")
                 stop(msg)
 
@@ -83,15 +81,16 @@ SSH = R6::R6Class("SSH",
         finalize = function(quiet = self$workers_running == 0) {
             #TODO: should we handle this with PROXY_CMD for break (and finalize if req'd)??
             if (private$ssh_proxy_running) {
-                rzmq::send.socket(private$proxy_socket,
-                      data=list(id="PROXY_STOP", finalize=!private$is_cleaned_up))
+                private$zmq$send(
+                    list(id="PROXY_STOP", finalize=!private$is_cleaned_up),
+                    "proxy"
+                )
                 private$ssh_proxy_running = FALSE
             }
         }
     ),
 
 	private = list(
-        proxy_socket = NULL,
         ssh_proxy_running = TRUE,
 
         fill_options = function(ssh_host, ...) {
@@ -101,7 +100,8 @@ SSH = R6::R6Class("SSH",
             #TODO: let user define ports in private$defaults here and respect them
             remote = sample(50000:55000, 2)
             values$ssh_host = ssh_host
-            values$local_port = bind_avail(private$proxy_socket, 11000:13000)
+            bound = private$zmq$listen(sid="proxy")
+            values$local_port = sub(".*:", "", bound)
             values$ctl_port = remote[1]
             values$job_port = remote[2]
             values$fwd_port = private$port