+++ tests/testthat/test-6-ssh_proxy.r
@@ -1,52 +1,46 @@
 context("proxy")
 
-has_localhost = has_connectivity("localhost")
+has_localhost = has_connectivity("127.0.0.1")
 
 test_that("control flow between proxy and master", {
     skip_if_not(has_localhost)
     skip_on_os("windows")
 
-    # prerequesites
-    context = rzmq::init.context()
-    socket = rzmq::init.socket(context, "ZMQ_REP")
-    port = bind_avail(socket, 50000:55000)
+    zmq = ZeroMQ$new()
+    port_ctl = as.integer(sub(".*:", "", zmq$listen())) #todo: sure this is always integer?
+    port_job = as.integer(sub(".*:", "", zmq$listen(sid="job")))
     common_data = list(id="DO_SETUP", fun = function(x) x*2,
             const=list(), export=list(), seed=1)
-    p = parallel::mcparallel(ssh_proxy(port, port, 'multicore'))
+    p = parallel::mcparallel(ssh_proxy(port_ctl, port_job, 'multicore'))
     on.exit(tools::pskill(p$pid, tools::SIGKILL))
 
     # startup
-    msg = recv(p, socket)
+    msg = zmq$receive()
     expect_equal(msg$id, "PROXY_UP")
+    worker_url = msg$worker_url
 
-    send(socket, common_data)
-    msg = recv(p, socket)
+    zmq$send(common_data)
+    msg = zmq$receive()
     expect_equal(msg$id, "PROXY_READY")
     expect_true("data_url" %in% names(msg))
     expect_true("token" %in% names(msg))
-    proxy = msg$data_url
-    token = msg$token
+#    token = msg$token
 
     # command execution
     cmd = quote(Sys.getpid())
-    send(socket, list(id="PROXY_CMD", exec=cmd))
-    msg = recv(p, socket)
+    zmq$send(list(id="PROXY_CMD", exec=cmd))
+    msg = zmq$receive()
     expect_equal(msg$id, "PROXY_CMD")
     expect_equal(msg$reply, p$pid)
 
-    # common data
-    worker = rzmq::init.socket(context, "ZMQ_REQ")
-    rzmq::connect.socket(worker, proxy)
-
-    send(worker, list(id="WORKER_READY"))
-    msg = recv(p, worker)
-    testthat::expect_equal(msg$id, "DO_SETUP")
-    testthat::expect_equal(msg$token, token)
-    testthat::expect_equal(msg[names(common_data)], common_data)
+    # start up worker
+    zmq$connect(worker_url, sid="worker")
+    zmq$send(list(id="WORKER_READY"), sid="worker")
+    msg = zmq$receive(sid="job")
+    testthat::expect_equal(msg$id, "WORKER_READY")
 
     # shutdown
-    msg = list(id = "PROXY_STOP")
-    send(socket, msg)
+    zmq$send(list(id = "PROXY_STOP"))
     collect = suppressWarnings(parallel::mccollect(p))
     expect_equal(as.integer(names(collect)), p$pid)
     on.exit(NULL)
@@ -56,9 +50,7 @@ test_that("full SSH connection", {
     skip_on_cran()
     skip_on_os("windows")
     skip_if_not(has_localhost)
-    skip_if_not(has_ssh_cmq("localhost"))
-    skip_if_not(identical(Sys.getenv("TRAVIS"), "true"),
-                message="this test runs on travis only")
+    skip_if_not(has_ssh_cmq("127.0.0.1"))
 
     # 'LOCAL' mode (default) will not set up required sockets
     # 'SSH' mode would lead to circular connections
@@ -69,7 +61,7 @@ test_that("full SSH connection", {
 
     options(clustermq.template = "SSH")
     w = workers(n_jobs=1, qsys_id="ssh", reuse=FALSE,
-                ssh_host="localhost", node="localhost")
+                ssh_host="127.0.0.1", addr="tcp://127.0.0.1:*")
     result = Q(identity, 42, n_jobs=1, timeout=10L, workers=w)
     expect_equal(result, list(42))
 })